Move CheckpointStorage and ReplayStorage into tlm_adjoint/checkpointing.py
